Updated directory structure and renamed files for ease of use
