+++ 2. Set-JAVNfo.ps1
