+++ emby_thumbs/3. Set-EmbyActorThumbs.ps1
