+++ settings_sort_jav.ini
@@ -104,7 +104,7 @@ keep-metadata-txt=true
 # Make sure that no other non-cover jpg files are in this directory or they may be cropped unintentionally
 scraped-covers-path=C:\Sorted
 
-# Settings for Get-R18ThumbUrl.ps1 ################################################################################
+# Settings for Get-R18ThumbUrls.ps1 ################################################################################
 
 # Page on https://www.r18.com/videos/vod/movies/actress/letter=a/sort=popular/ to start scraping actress links from
 r18-start-page=1
